Refactor transaction testing
